Kang

+++ src/main/java/com/heartsignal/dev/service/domain/rds/MeetingChatRoomService.java
@@ -24,12 +24,12 @@ public class MeetingChatRoomService {
     public Long makeMeetingChatRoom(Team team1, Team team2) {
         StoredProcedureQuery storedProcedure = entityManager.createStoredProcedureQuery("createMeeting");
 
-        storedProcedure.registerStoredProcedureParameter("v_team1_id", Integer.class, ParameterMode.IN);
-        storedProcedure.registerStoredProcedureParameter("v_team2_id", Integer.class, ParameterMode.IN);
+        storedProcedure.registerStoredProcedureParameter("v_team1_id", Long.class, ParameterMode.IN);
+        storedProcedure.registerStoredProcedureParameter("v_team2_id", Long.class, ParameterMode.IN);
         storedProcedure.setParameter("v_team1_id", team1.getId());
         storedProcedure.setParameter("v_team2_id", team2.getId());
 
-        storedProcedure.registerStoredProcedureParameter("o_meeting_id", Integer.class, ParameterMode.OUT);
+        storedProcedure.registerStoredProcedureParameter("o_meeting_id", Long.class, ParameterMode.OUT);
         storedProcedure.execute();
